feat: finished creating all crud methods

+++ src/columns/columns.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { ForbiddenException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { UserEntity } from 'src/users/entity/users.entity';
+import { AuthorizedUser } from 'src/types';
 import { FindManyOptions, FindOneOptions, Repository } from 'typeorm';
 import { CreateColumnsDto } from './dto/create-dto';
 import { UpdateColumnsDto } from './dto/update-dto';
@@ -13,14 +13,11 @@ export class ColumnsService {
     private readonly columnsRepository: Repository<ColumnEntity>,
   ) {}
 
-  async create(
-    authorizedUser: Omit<UserEntity, 'password'>,
-    dto: CreateColumnsDto,
-  ) {
+  async create(authorizedUser: AuthorizedUser, dto: CreateColumnsDto) {
     return await this.columnsRepository.save({ ...dto, owner: authorizedUser });
   }
 
-  async getOne(query: FindOneOptions) {
+  async findOne(query: FindOneOptions) {
     return await this.columnsRepository.findOne(query);
   }
 
@@ -28,11 +25,28 @@ export class ColumnsService {
     return await this.columnsRepository.find(query);
   }
 
-  async update(columnId: number, dto: UpdateColumnsDto) {
-    return await this.columnsRepository.update(columnId, dto);
+  async update(
+    authorizedUser: AuthorizedUser,
+    id: number,
+    dto: UpdateColumnsDto,
+  ) {
+    await this.checkingColumnsOwner(authorizedUser.id, id);
+    return await this.columnsRepository.update(id, dto);
   }
 
-  async remove(id: number) {
+  async remove(authorizedUser: AuthorizedUser, id: number) {
+    await this.checkingColumnsOwner(authorizedUser.id, id);
     return await this.columnsRepository.delete(id);
+  }
+
+  async checkingColumnsOwner(userId: number, columnId: number) {
+    const { owner } = await this.findOne({
+      where: { id: columnId },
+      relations: ['owner'],
+    });
+
+    if (owner.id !== userId) {
+      throw new ForbiddenException('У вас нет доступа');
+    }
   }
 }
